requirements.txt
dbfread==2.0.7
numpy==2.0.2
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
python-dotenv==1.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
tqdm==4.66.1